Password bubble: ManagePasswordsIconView is now a BubbleIconView.
This CL introduces a new browser command to open the Manage Passwords bubble,
and converts the ManagePasswordsIconView class into a subclass of
BubbleIconView, which uses the new command to control the bubble's state.
This allows us to more easily test the view and the UI controller, as each
object's job is now more clearly defined (and the view is now doing a good
deal less work), and to independently verify that the command is doing the
right thing.

After this CL, we'll (finally!) have something approaching reasonable test
coverage for the core of the views code.

----------------------------------------------------------------------------
This is a re-land of r267195, which was reverted due to Windows errors. That
was a reland of r266859, which was reverted due to LSAN errors. The
original review is https://codereview.chromium.org/246393004/.
----------------------------------------------------------------------------

BUG=365678
TBR=vabr@chromium.org,pkasting@chromium.org,cpu@chromium.org

Review URL: https://codereview.chromium.org/264713010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268159 4ff67af0-8c30-449e-8e8b-ad334ec8d88c