components: Do not include test code in wifi_component target.
wifi_component target is production code target and thus should not
include code that is intended for tests.

This changes makes it stop building test code in wifi_component, by
extracting and exposing the FakeWiFiService interface in the header
file and instantiating it directly where we used to call
WiFiService::CreateForTest().

This is the second part of what was discussed in
https://codereview.chromium.org/243343002 and started here
https://codereview.chromium.org/259543002/ - r266025

BUG=None
TEST=build and works as before, no regressions
R=mef@chromium.org,thestig@chromium.org

Review URL: https://codereview.chromium.org/264773017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268237 4ff67af0-8c30-449e-8e8b-ad334ec8d88c