Attempt Manager Refactor and Opt-In Pipeline Refactor
Refactored the attempt manager to use a more obvious naming scheme as well as
participate in promises for retries.

This also entailed reexamination of all attempt managers in Chrome Now.
Opt-in checking as a result became a first class state change citizen and is
no longer part of the card checking pipeline, fixing a few opt-in aggression
issues. With this change, we don't care how often a push message comes through
that may indicate that Google Now is enabled.

BUG=360649,362306

Review URL: https://codereview.chromium.org/248473003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268302 4ff67af0-8c30-449e-8e8b-ad334ec8d88c