Allow the PDF extension to be run in incognito mode.
This allows PDF to be run in incognito mode but changing the incognito manifest key to "split". This causes a new process to be loaded for the extension in incognito. This is a requirement for doing top level navigations to an extension in incognito mode as per https://code.google.com/p/chromium/codesearch#chromium/src/extensions/browser/extension_protocols.cc&rcl=1399166280&l=287

BUG=303491

Review URL: https://codereview.chromium.org/268963002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268308 4ff67af0-8c30-449e-8e8b-ad334ec8d88c