Use FakeServer-based invalidations for Sync tests
Note: this is a second go-round for this CL. It was originally committed
as r267422 but was reverted due to a compile error in some dead code.
The dead code was deleted in https://codereview.chromium.org/264773007/
 (r267686). See https://codereview.chromium.org/234113002/ for the
original CL.

This CL creates a new InvalidationService implementation,
FakeServerInvalidationService, to remove FakeServer-based tests'
dependency on the Python XMPP server.

Another major change is that FakeServer's HandleCommand is now executed
on the UI thread and locking has been removed from FakeServer itself.

BUG=323265

Review URL: https://codereview.chromium.org/267723012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268311 4ff67af0-8c30-449e-8e8b-ad334ec8d88c