ThreadWatcher - On Android, don't generate the crash dump if StartupTimebomb goes off.
When StartupTimebomb goes off, collect the duration stats (difference
between when the timebomb went off and when it is started). We collect
this duration in terms of clock time, montonic clock time and thread CPU
time.

We wanted to see when the browser process is stopped, does
TimeTicks::Now() account for the time the browser process was stopped or
not. If the TimeTicks doesn't account this time properly, then we could
have false positive jank related crashes.

BUG=366699
R=jar@chromium.org

Review URL: https://codereview.chromium.org/269763002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268338 4ff67af0-8c30-449e-8e8b-ad334ec8d88c