Don't consider shutdown acknowledgement for external extensions
External extensions display a warning, and, on shutdown, we close that warning
and treat it as an "abort". This has the effect of uninstalling the extension -
this is bad, because a) the user won't expect it, and b) we're in the process of
shutting down and shouldn't be doing something like uninstalling. Only treat an
abort as an uninstall if it's user initiated.
BUG=370222

Review URL: https://codereview.chromium.org/261713003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268947 4ff67af0-8c30-449e-8e8b-ad334ec8d88c