Add browser_test for extension app API with missing schema
This is a simplified version of api_test/stubs. It touches all the parts
of each chrome.foo.bar.baz API path and ensures they are defined. This
catches cases where a developer adding an API to _api_features.json but
did not provide a schema for them.

BUG=369318
TEST=added browser_tests ExtensionApiTest.StubsApp
R=kalman@chromium.org

Review URL: https://codereview.chromium.org/264923009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268949 4ff67af0-8c30-449e-8e8b-ad334ec8d88c