Make quotes to spaces in logs passed to PasswordManagerInternalsUI
Currently, logs passed to PasswordManagerInternalsUI are a JSON string, which gets escaped for HTML, and then displayed in JavaScript by assigning to a DIV element's .innerText

The problem is with the quote character ("), which appears frequently in the JSON string, and gets expanded to an ugly &quote;, making the logs very hard to read.

This CL adds one more step before the escaping -- replacing the quote (") character with a space.

Another option would be to rely on the net::EscapeForHTML function and assign the result to .innerHTML instead of .innerText, but that was vetoed in https://codereview.chromium.org/235623002/diff/20001/chrome/browser/resources/password_manager_internals/password_manager_internals.js#newcode12.

BUG=347927

Review URL: https://codereview.chromium.org/267983003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@268954 4ff67af0-8c30-449e-8e8b-ad334ec8d88c