Beginning of support for extension content verification
The basic idea is that the webstore will provide signed expected hashes
of file content that can be checked during runtime in the browser to
detect corruption due to disk errors or malware.

This CL has a lot of the high-level pieces, with several of the details
left out for subsequent CLs to make this one more easily digestible.

The design is that there is a ContentVerifier for each BrowserContext
which can be used anywhere we read from files inside an extension. It
vends out ContentVerifyJob's, which need to be informed of the bytes
read from each file, and will know how to check those against a set of
expected block hashes. If the job detects contents that don't match what
was expected, it will notify the verifier.

BUG=369895
R=rvargas@chromium.org, yoz@chromium.org

Review URL: https://codereview.chromium.org/266963003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@269108 4ff67af0-8c30-449e-8e8b-ad334ec8d88c