PasswordManagerInternalsUI: Postpone logging after WebContents loads
Currently, logs sent to the internals page controller are passed to the internals
page JavaScript code immediately. This is mostly OK, except for a short time between
the construction of the UI controller, and loading the internals page.

Once multiple internals pages can be opened (this is being addressed by
https://codereview.chromium.org/262583007/ and a follow-up https://codereview.chromium.org/269513003/), the above
issue becomes more apparent: the second internals page receives the content of
the first one on construction, so that both pages display the same logs.
However, the construction happens before the page itself finishes loading.
As a result, the old logs are never displayed on new internals pages.

This CL makes the internals page controller a WebContentsObserver. The controller
accumulates logs until the WebContents singnalises loading finished, at which point
the controller first sends the logs for display.

The CL also contains a test for this, re-creating the issue described above with
two tabs open. However, that test must currently be disabled, until it becomes
possible to actually open two tabs with an internals page.

BUG=347927

Review URL: https://codereview.chromium.org/277493002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@269285 4ff67af0-8c30-449e-8e8b-ad334ec8d88c