Reland: Support multiple sign-in for extension updates.
This changes the ExtensionDownloader to attempt multiple
sequential queries for CRX downloads, iterating over incremental
values for the authuser query parameter up to a max of 10.

This is a re-landing of http://codereview.chromium.org/279453002/
which was reverted due to a use-after-free bug in the unit tests.

BUG=370964
R=asargent@chromium.org

Review URL: https://codereview.chromium.org/274263002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@269666 4ff67af0-8c30-449e-8e8b-ad334ec8d88c