Implement new bookmark app icon generation.
This CL changes the generation of bookmark apps to create an icon that
consists of the favicon's dominant color as a background with the first
letter of the app's domain painted over it.

The letter is either black or white depending on the luminance of the
background.

This CL also turns on generation of 2x icons for 32 and 48 px icons.

BUG=315632

Review URL: https://codereview.chromium.org/246333006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@269745 4ff67af0-8c30-449e-8e8b-ad334ec8d88c