Add a skeleton Start Page to the experimental app list.
This CL adds a basic version of a start page to the experimental app list.
This page contains a WebView that displays the Google Logo and a grey bar
which are currently only placeholders.

This CL also modifies the StartPage WebUI to have the google logo.

BUG=349727

Review URL: https://codereview.chromium.org/186483004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@269752 4ff67af0-8c30-449e-8e8b-ad334ec8d88c