Remove RenderWidget::SetBackground, change IPC to pass a bool.
The ViewMsg_SetBackground message currently passes an SkBitmap that
signals if the RenderView should have an opaque background or not. We
can use a bool instead to convey this data.

Rename the ViewMsg_SetBackground to ViewMsg_SetBackgroundOpaque as well
as the Set/GetBackground methods on the various RenderWidget/View
classes.

RenderWidget itself no longer needs to know about the background opaque
changes.

R=kenrb, piman
BUG=362164

Review URL: https://codereview.chromium.org/274453002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@269953 4ff67af0-8c30-449e-8e8b-ad334ec8d88c