Fix a bug where two loaded keywords clash after a GoogleBaseURL change.
Repros when a keyword exists for google.ca, in addition to a pre-pop for google.com, and the user is in Canada. The google.com is rewritten to google.ca and the latter should be uniquified to google.ca_. But it was not working as intended.

NOTREECHECKS=True
NOTRY=True
BUG=365762

Review URL: https://codereview.chromium.org/287563004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@270258 4ff67af0-8c30-449e-8e8b-ad334ec8d88c