Extract logs persistence code to metrics component.
This extracts the code for storing persisting a list of unsent logs in
preference, and refactors the code that depends on it.

The SerializedLog class is changed to a private type LogHashPair in PersistedLogs.

The PersistedLogs class replaces the MetricsLogSerializer class, and additionally
takes ownership of the list of logs being managed.

With the current implementation, when a log is staged, it is not removed from
the PersistedLogs object, and is instead removed when the staged log is
discarded.  This removes the need for provisionally storing the staged log if
it is currently being uploaded when shutting down.

BUG=372911

Review URL: https://codereview.chromium.org/178833004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@270557 4ff67af0-8c30-449e-8e8b-ad334ec8d88c