Handle X11 Mod3Mask for Caps Lock remapping.
Mod3Mask is set on X events when the Caps Lock key is down, but, if
Caps Lock is remapped, the X CapsLock flag is NOT set, because pressing
the key does not invoke caps lock from X's point of view. So, have the
kModifierRemappings[] table uses EF_MOD3_DOWN for the Caps Lock remapping.

BUG=372796
R=derat@chromium.org
TEST=manual on device

Review URL: https://codereview.chromium.org/277443004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@270704 4ff67af0-8c30-449e-8e8b-ad334ec8d88c