Makes construction of (ash)RemoteWindowTreeHostWin explicit
I need to do this as I need to pass in state to the constructor and I
can't do that with lazy construction. This makes creation a little
saner anyway.

I'm not happy about the static setting the HWND. I'll see if I can
clean that up later.

BUG=none
TEST=none
R=ananta@chromium.org

Review URL: https://codereview.chromium.org/277753002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@270747 4ff67af0-8c30-449e-8e8b-ad334ec8d88c