Fix gesture event handling in MenuButton
In MenuButton, when Activate() function returns true, it means the
button has not shown a drop-down menu, so the gesture event handler
should continue handling the event passing it up to the base class
handler. On the other hand, if it returns false, the drop-down menu is
shown and no more processing is needed by the button.
Also, touch_down_context_ in HWNDMessageHandler is changed from boolean
to integer in order to account for multiple touches happening right
after each other.

BUG=349881

Review URL: https://codereview.chromium.org/271863006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@270756 4ff67af0-8c30-449e-8e8b-ad334ec8d88c