[OriginChip] Re-enable the chip if the text changes during an Update().
Note that the outcome now in the case where you click the chip and then the back button is that the Omnibox retains focus but the chip is shown.  We normally don't allow this but it's analogous to the non-origin-chip behavior in this case.

BUG=371684

Review URL: https://codereview.chromium.org/290623002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@271177 4ff67af0-8c30-449e-8e8b-ad334ec8d88c