drive::DownloadHandler should query DownloadHistory.
The drive::DownloadHandler was using the assumption that all
non-IN_PROGRESS downloads that are created were restored from history.
This is not a safe assumption.

Instead query the DownloadHistory to determine whether a new download
was restored from history or not.

Depends on https://codereview.chromium.org/230103002/ which adds a
WasDownloadRestoredFromHistory method to DownloadHistory.

BUG=7648

Review URL: https://codereview.chromium.org/230133002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@271595 4ff67af0-8c30-449e-8e8b-ad334ec8d88c