Create GoogleURLTrackerClient interface and //chrome implementation
GoogleURLTrackerClient is the interface via which GoogleURLTracker will talk to
its embedder. This CL creates that interface as well as
ChromeGoogleURLTrackerClient, the //chrome implementation. The interface and
implementation are carved out of GoogleURLTrackerNavigationTracker(Impl),
specifically the parts that are not per-tab. A followup CL will modify
GoogleURLTrackerNavigationTracker to the driver interface for
GoogleURLTracker.

BUG=373220

Review URL: https://codereview.chromium.org/285193002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@271638 4ff67af0-8c30-449e-8e8b-ad334ec8d88c