Add 4k HW decode override flag and support larger bitstream buffers in V4L2VDA.
Add a flag to override 4k hardware video decode capability detection in GVD
if needed. This is for platforms that currently do not give us a proper way
to detect it.

Also bump input bistream size in V4L2VDA to accommodate potentially
larger streams in higher resolutions. We cannot estimate input buffer size
based on the stream, as we do not currently pass video config info down to
the VDAs, and the bitstream buffers are needed in order to parse the
stream to get that information.

BUG=373717
TEST=4k video playback

Review URL: https://codereview.chromium.org/285343003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@271653 4ff67af0-8c30-449e-8e8b-ad334ec8d88c