<webview>: Move NewWindow API to chrome
This CL moves the New Window API and navigation to chrome.

This CL removes a lot of unnecessary content APIs for BrowserPlugin now that permissions no longer live in content.

Up next, in a separate CL will be a huge cleanup that removes all WebContentsDelegate code out of BrowserPluginGuest and into WebViewGuest. This should significantly reduce the size of the BrowserPlugin content API.

BUG=364141, 330264
TBR=kenrb@chromium.org for trivial browser_plugin_messages.h change.

Review URL: https://codereview.chromium.org/272573005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@271703 4ff67af0-8c30-449e-8e8b-ad334ec8d88c