Allow extension crash notifications to replace each other.
Previously they did not set an origin or replace_id, causing updates
to be turned into adds in the notification UI manager, where it would
DCHECK because the id matched that of an existing notification.

This patch adds a replace_id and origin so that updates can happen
smoothly.

BUG=360591

Review URL: https://codereview.chromium.org/294473002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@272006 4ff67af0-8c30-449e-8e8b-ad334ec8d88c