Turn GoogleURLTrackerNavigationHelper(Impl) into a per-tab object.
The goal of this CL is to eliminate the dependence of
GoogleURLTracker(MapEntry) on NavigationController. To accomplish this goal,
GoogleURLTrackerNavigationHelper is turned into a conceptually per-tab
interface. GoogleURLTracker::OnNavigationPending() now takes in the
GoogleURLTrackerNavigationHelper with which the navigation is associated rather
than the associated NavigationController; GoogleURLTracker performs per-tab
actions by calling the navigation helper associated with the tab.

A followup CL will turn GoogleURLTrackerNavigationHelper(Impl) into the
GoogleURLTrackerDriver interface and ContentURLTrackerDriver implementation.

BUG=373230

Review URL: https://codereview.chromium.org/283413002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@272094 4ff67af0-8c30-449e-8e8b-ad334ec8d88c