Session restore shouldn't care about profile home pages.
The logic where we treated an empty URL as home page is outdated.

This CL contains the following additional fixes:

1) Disallow using empty URLs to denote anything (home page, new tab
page...) and force the callers of chrome::Navigate to specify which URL they
want to open. Using empty URLs was hiding bugs (see below).

2) Fixed StartupBrowserCreatorTest.UpdateWithTwoProfiles and
ProfilesWithoutPagesNotLaunched so that they don't work by accident (about:blank
used to be the home page - it wasn't restoring the previous session, but
launching the home page).

3) There was some code passing GURL("new_tab_page") around, and
comparing against that. But that's an invalid URL, so GURL will just make it
empty, and the result is not what is expected. (E.g,. GURL("foo") ==
GURL("new_tab_page")).

4) Fixed other places which were passing GURL("something_invalid") around and
pretending it's something meaningful. It was just a coincidence that nothing was
broken.

BUG=371852
R=sky@chromium.org

Review URL: https://codereview.chromium.org/292713003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@272182 4ff67af0-8c30-449e-8e8b-ad334ec8d88c