sync: Implement fake server reflection blocking
Implements reflection blocking for the in-process fake sync server.

The notifications delivery system used in the real world will not
deliver notifications of new or modified items to the client that
originated that change.  This feature is known by the names
"reflection blocking" or "squelching".

Most of out integration tests expect reflection blocking to be
disabled, so we can't turn this on in all tests.  The default
is to leave it off.  Tests that want to enable reflection blocking
must override "TestUsesSelfNotifications" in their test framework
class.

BUG=323265

Review URL: https://codereview.chromium.org/296743005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@272321 4ff67af0-8c30-449e-8e8b-ad334ec8d88c