[Win] Draw the circular avatar programmatically rather than with a frame.
On Linux, the OS theme can change, which affects the default background
colour of the avatar bubble. This means that using a frame overlay to
display the profile avatar as an image is not actually feasible.

This also adds a small change to the ClipPath function of gfx::Canvas,
to allow controlling the anti-aliasing option.

BUG=372721
TEST=Start Chrome with --new-profile-management. Click on the avatar button. The avatar should still be displayed as a circle.

Review URL: https://codereview.chromium.org/290333012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@272854 4ff67af0-8c30-449e-8e8b-ad334ec8d88c