Implement special path and mimetype check in file_manager/filesystem_api_util.
Previously, only the fixed path for Drive mount point was treated as "special",
meaning that it does not correspond to a native local path. With this CL, all
mount points with non-"NativeLocal" file system types are handled as special.

Since mimetype metadata is only available for Drive as of now, on other special
paths, the GetSpecialPathMimeType function returns success with empty mimetype.

BUG=367028

Review URL: https://codereview.chromium.org/304503007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273560 4ff67af0-8c30-449e-8e8b-ad334ec8d88c