Port aria_util_test.js to run as a WebUI test.
This takes one existing ChromeVox unit test and ports it to
run as a WebUI test in Chrome. It's a WebUI test because it
needs a full DOM to run in, even though it's otherwise most
similar to a unit test.

Only one change to the WebUI test system was needed - to
allow paths relative to the source root for library files for a
test. The other changes improve the error messages when
building, and add filenames to stack traces.

BUG=371692
NOTRY=true

Review URL: https://codereview.chromium.org/292313004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273577 4ff67af0-8c30-449e-8e8b-ad334ec8d88c