Implements RTP header dumping.
WebRtcRtpDumpHandler implements RTP header dump creation. It's owned by WebRtcLoggingHandlerHost and receives RTP packet callbacks through P2PSocketHost.

WebRtcRtpDumpWriter is owned by WebRtcDumpHandler and writes the RTP header into a memory buffer. When the in-memory buffer is full, it compresses the data and writes it to the disk.

After the dumping is stopped, WebRtcLoggingHandlerHost calls ReleaseDump to get the completed dump as a .gz file and uploads it.


BUG=363459

Review URL: https://codereview.chromium.org/264793017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273745 4ff67af0-8c30-449e-8e8b-ad334ec8d88c