Add option to install an ephemeral app to ChromeOS shelf context menu
This is the ChromeOS implementation for adding an option to install an
ephemeral app in the shelf context menu. Includes some refactoring of
the WebstoreStandaloneInstallers to reuse common code.

BUG=339253
TEST=browser_tests (LauncherPlatformAppBrowserTest.InstallEphemeralApp)
Manual: Enable the ephemeral apps experimental feature and launch an
ephemeral app. Right-click on the app's icon in the shelf. There should
be an "Install app" item, which invoked, will install the app and make
it visible in the app launcher and NTP.

Review URL: https://codereview.chromium.org/298303002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273749 4ff67af0-8c30-449e-8e8b-ad334ec8d88c