Move ChromeOS hardware class init out of MetricsService.
This CL moves the initialization of the hardware class on ChromeOS out of
MetricsService and into ChromeOSMetricsProvider via ChromeMetricsServiceClient.
MetricsService now pass |OnInitTaskGotHardwareClass()| as the callback to
ChromeMetricsServiceClient::StartGatheringMetrics().
ChromeMetricsServiceClient::StartGatheringMetrics() posts a task to the FILE
thread calling ChromeMetricsServiceClient::InitTaskGetHardwareClass(). The
latter function calls into ChromeOSMetricsProvider when on ChromeOS and
directly calls the callback passed to it on other platforms.

A followup CL will move the rest of the embedder-specific initial metrics
gathering flow out of MetricsService.

BUG=375776

Review URL: https://codereview.chromium.org/301633006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273820 4ff67af0-8c30-449e-8e8b-ad334ec8d88c