Use .volume_label property for notifying MTP volume labels.
Previously we've abused .volume_id for the purpose, but now we have a
dedicated field for the volume name.

Now the .volume_id field can be arbitrary human unreadable string,
but I'd like to keep it as in the current form since there's also no
other reason to make it cryptic.

BUG=none

Review URL: https://codereview.chromium.org/305053004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273826 4ff67af0-8c30-449e-8e8b-ad334ec8d88c