Omnibox: HistoryQuick Provider: Index Farther Into URL and Title
Record word break information for the first 200 characters in a URL or
title, not the first 50 characters.  This is important because we
generally drop URL suggestions that we think do not match at a word
break.

Tested locally on a longer URL from the bug.

TBR=sky
(because this change is trivial)

BUG=291294

Review URL: https://codereview.chromium.org/306963002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273837 4ff67af0-8c30-449e-8e8b-ad334ec8d88c