Remove ViewMsg_SetZoomLevel
In an effort to simplify zoom, this CL removes ViewMsg_SetZoomLevel and plumbs zoom on WebContentsImpl directly through to HostZoomMap.

At present calls to WebContents::SetZoomLevel() invoke up to three cross-process IPCs, and introduce a synchronization requirement for any caller that needs to block until the results of the operation are visible via WebContents::GetZoomLevel().

This CL simplifies this down to one IPC and removal of the
synchronization requirement. It adds a new IPC message to signal changes in a view's ability to be zoomed, and removes an ViewMsg_SetZoomLevel that was used only by SetZoomLevel().

BUG=none

Review URL: https://codereview.chromium.org/287093002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273854 4ff67af0-8c30-449e-8e8b-ad334ec8d88c