Remove flag cloud-policy-invalidation-delay.
The cloud policy server is now controlling the amount of delay through policy, so the flag is no longer needed. This change also lowers the default amount of delay (in absence of policy) to a more reasonable value based on observations of traffic to the cloud policy server.

BUG=379010

Review URL: https://codereview.chromium.org/305073002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273913 4ff67af0-8c30-449e-8e8b-ad334ec8d88c