Browser Plugin: Remove Zoom
With the zoom refactor in flight: https://codereview.chromium.org/287093002/, zoom
plumbing for BrowserPlugin in the content layer is now entirely dead code. We need to revisit zoom in
the chrome layer for <webview> so that the embedder's zoom is propagated to the guests.

This will be done at a later time when we need to get <webview> in WebUI working again.

BUG=330264
TBR=jam@chromium.org for trivial browser_plugin_guest_delegate.h code deletion.

Review URL: https://codereview.chromium.org/303423002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@273959 4ff67af0-8c30-449e-8e8b-ad334ec8d88c