sync: Add UI code for per type debug counters
Adds a new tab to the about:sync page to display the latest per-type
counters.  These counters will eventually replace the StatusControllers
that contain much of the same information.

This includes some changes to the C++ SyncInternalsMessageHandler to
register to observe these counters as well as some new WebUI to display
them.

There is a known bug in this code.  The StausCounter update logic has
not been implemented yet, so the item counts will always be zero.  This
will be fixed in a future CL.

BUG=349301

Review URL: https://codereview.chromium.org/290023003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274348 4ff67af0-8c30-449e-8e8b-ad334ec8d88c