Animate window control changes in TouchView
Add animation for the hiding and showing of the resize button. Add an animation
for sliding the minimize button into the position of the resize button. Delay
changing the size of FrameCaptionButtonContainerView to account for conflicting
layout changes caused by the transition to TouchView.

TEST=FrameCaptionButtonContainerViewTest
TEST=CustomFrameViewAshTest
TES=BrowserNonClientFrameViewAshTest
BUG=363717

Review URL: https://codereview.chromium.org/271913002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274371 4ff67af0-8c30-449e-8e8b-ad334ec8d88c