Add ability to constrain dominant color selection to a HSL range.
This CL adds the ability to supply CalculateKMeanColorOfPNG with a lower
and upper HSL bound which will be used to select the dominant color.

This CL also adds CalculateKMeanColorOfPNG with only a PNG param as a
convenience method, using sane defaults and a utility function
IsWithinHSLRange for checking whether a given HSL is within the given
HSL range.

BUG=376613
TBR=benwells@chromium.org, estade@chromium.org

Review URL: https://codereview.chromium.org/289283004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274500 4ff67af0-8c30-449e-8e8b-ad334ec8d88c