Remove unnecessary copies in CRLSet parsing.
This showed up in Chrome Traces covering the startup of Chrome for Android.
CRLSet parsing was taking 70ms of CPU on Nexus 4 (although this is happening on
the File thread).
That cost is now down to 18ms. There are certainly further ways to decrease
this even more though.

BUG=376793

Review URL: https://codereview.chromium.org/302643003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274607 4ff67af0-8c30-449e-8e8b-ad334ec8d88c