linux_aura: Theoretical crash for full screen transition.
The only way this crash could occur is if we're trying to layout the
titlebar, the delegate requests showing the icon, but there isn't an
icon view. In such cases, we shouldn't try to set the bounds on a NULL
view.

(This fix is theoretical; I wasn't able to repro the bug.)

BUG=376313

Review URL: https://codereview.chromium.org/310193002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274749 4ff67af0-8c30-449e-8e8b-ad334ec8d88c