Fix a crash when chooseDesktopMedia is called with no target tab
DesktopMediaPicker was getting created with web_contents returned by
web_contents(), which is part of WebContentsObserver interface,
and which does not get set when the target tab is not specified.

Also, fix handling of a case when the method is called when there are
no web contents to attach the picker dialog to.

BUG=375576,364050
TEST=from a extension with desktopCapture permission:
chrome.desktopCapture.chooseDesktopMedia(['screen'], undefined, function())
does not crash

Review URL: https://codereview.chromium.org/307453010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274771 4ff67af0-8c30-449e-8e8b-ad334ec8d88c