Refactor app list so AppsGridView owns the PaginationModel.
Previously, PaginationModel was created by platform-specific code
(either AppListController or AppListShower) and passed down through 6
layers to AppsGridView. Now, AppsGridView creates its own
PaginationModel, and many of the layers don't need to know about it.

Some classes do need to access the PaginationModel (most of these 
dependencies can probably be broken later). These now pull the 
PaginationModel from the AppsGridView instead of caching it on the way 
down. 

On ChromeOS, the lifetime of the PaginationModel is now tied to the 
visibility of the app list (was previously the lifetime of the process). 
To get around this, AppListController must now manually remember the 
current page, passing it to new PaginationModels as they are created, as 
well as adding itself as an observer. This is still a net win, as the 
ownership is in the right place, and several classes no longer depend on 
PaginationModel, with more dependency cuts possible in the future. 

BUG=377626

Review URL: https://codereview.chromium.org/302803002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274785 4ff67af0-8c30-449e-8e8b-ad334ec8d88c