Fix crash when calling getMsg with no options.
(This is cherry-picking a single bug fix from ChromeVox upstream.)

The repro is this html:
<div tabIndex=0 role="slider" aria-valuenow="20"></div>

The bug happens when cvox.BrailleUtil.getState calls
cvox.ChromeVox.msgs.getMsg to see if the "_brl" translation exists,
and passes it 'aria_value_now_brl' in this particular case,
whose value is just "$1". If you don't pass it an argument, you
get '' back, and that throws an error because a message should never
return the empty string.

BUG=379296
R=plundblad@chromium.org

Review URL: https://codereview.chromium.org/302063005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274815 4ff67af0-8c30-449e-8e8b-ad334ec8d88c