Add DCHECK to MaskedViewTargeter::GetHitTestMask() overrides
Add DCHECK_EQ(view, masked_view()) to the overrides of
MaskedViewTargeter::GetHitTestMask(). This is to ensure
the function is only called with |view| corresponding
to the view on which the targeter is installed.

This is nice to have because GetHitTestMask()
is called directly from a few places such as
overrides of View::GetHitTestMask() and (soon)
from within TabCloseButtonTargeter; see issue
377557.

BUG=377527
TEST=none

Review URL: https://codereview.chromium.org/308003019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@274852 4ff67af0-8c30-449e-8e8b-ad334ec8d88c